Implement initial version of text justification, new tests, and fix bugs with right and center alignment.

+++ engine/src/flutter/src/paragraph_style.h
@@ -32,7 +32,7 @@ class ParagraphStyle {
   FontStyle font_style = FontStyle::normal;
   std::string font_family = "";
   double font_size = 14;
-  int max_lines = 1;
+  size_t max_lines = 1;
   double line_height = 1.0;
   std::string ellipsis;
 };
